Move __version__ to separate file to fix circular import
