+++ haystack/preview/components/fetchers/link_content.py
@@ -8,9 +8,9 @@ from requests import Response
 from requests.exceptions import HTTPError
 from tenacity import RetryCallState, retry, retry_if_exception_type, stop_after_attempt, wait_exponential
 
-from haystack.preview import __version__
 from haystack.preview import component, default_from_dict, default_to_dict
 from haystack.preview.dataclasses import ByteStream
+from haystack.preview.version import __version__
 
 logger = logging.getLogger(__name__)
